pyproject.toml
[project]
name = "ragagent2"
version = "0.1.0"
description = "A Graph RAG system for technical documents with hybrid retrieval"
readme = "README.md"
requires-python = ">=3.12.1"
authors = [
    {name = "Jason Helland"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # PDF Processing
    "docling>=2.0.0",
    "pypdf>=3.0.0",
    "opencv-python-headless>=4.9.0",
    # NLP and Entity Extraction
    "spacy>=3.7.0",
    # LLM Integration
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    # Graph Database
    "neo4j>=5.15.0",
    # Vector Database
    "qdrant-client>=1.7.0",
    "fastembed>=0.2.0",
    # Embeddings
    "sentence-transformers>=2.3.0",
    # Normalization and Matching
    "numpy>=1.24.0",
    "rapidfuzz>=3.8.0",
    # Data Validation and Configuration
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "PyYAML>=6.0",
    # CLI and Utilities
    "loguru>=0.7.2",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "textual>=0.47.0",
    # Testing
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "spacy-transformers>=1.3.9",
    "pip>=25.3",
    "en-core-web-lg",
    "en-core-web-sm",
    "ruff>=0.14.9",
    "black>=25.12.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.12.0",
    "ruff>=0.1.9",
    "ipython>=8.18.0",
    "jupyter>=1.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-recording>=0.13.0",
]

[project.scripts]
ragagent-ingest = "scripts.ingest_documents:main"
ragagent-discover = "scripts.run_discovery:main"
ragagent-review = "scripts.review_entities:main"
ragagent-review-interactive = "scripts.review_entities_interactive:main"
ragagent-setup = "scripts.setup_databases:main"
ragagent-query = "scripts.query_system:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# The project’s import root is the `src` package directory (this repo uses `from src...` imports).
# Hatchling cannot infer that from the project name `ragagent2`, so we configure it explicitly.
# We also include the `scripts` package so console entry points can import CLI helpers.
packages = ["src", "scripts"]

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    # TODO: migrate codebase to modern PEP 585/604 typing; enabling these now yields repo-wide lint noise
    "UP006",  # non-pep585-annotation
    "UP035",  # deprecated-import
    "UP045",  # non-pep604-annotation-optional
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html"
markers = [
    "e2e: End-to-end tests"
]

[tool.uv.sources]
en-core-web-lg = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl" }
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
